+++ src/helpers/getTaxFromShippingOptions.js
@@ -1,8 +1,8 @@
 const getTaxAndDutyFromShippingOptions = shippingOptions =>
 	shippingOptions.reduce(
 		(acc, { tax, duty }) => ({
-			tax: parseInt(tax || 0) + acc.tax,
-			duty: parseInt(duty || 0) + acc.duty
+			tax: parseFloat(tax || 0) + acc.tax,
+			duty: parseFloat(duty || 0) + acc.duty
 		}),
 		{
 			tax: 0,
